# Improve perf of CharEnumerator