The C# compiler lowers a foreach over a string to a for loop equivalent, so this is only relevant when the string is used as an `IEnumerable<char>`.

```C#
private string _s = new string('s', 100);

[Benchmark]
public int Iterate()
{
    int sum = 0;
    foreach (char c in (IEnumerable<char>)_s) sum += c;
    return sum;
}
```

|  Method |         Toolchain |     Mean |   Error |  StdDev | Ratio |
|-------- |------------------ |---------:|--------:|--------:|------:|
| Iterate | \main\corerun.exe | 301.4 ns | 2.76 ns | 2.58 ns |  1.00 |
| Iterate |   \pr\corerun.exe | 235.7 ns | 1.13 ns | 1.06 ns |  0.78 |